requirements.txt
starlette
bcrypt
pycryptodome
requests